requirements.txt
Django>=5.2,<6.0
gunicorn>=22.0,<23.0
tensorflow-cpu>=2.16,<3.0
opencv-python-headless>=4.9,<5.0
numpy>=1.26,<2.0
Pillow>=10.0,<11.0
whitenoise>=6.5,<7.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
